# TodoList/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/todoList": {
            "post": {
                "description": "Create a Todo List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todoList"
                ],
                "summary": "Create todoList",
                "parameters": [
                    {
                        "description": "Create todoList",
                        "name": "todoList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TodoListDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TodoListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            }
        },
        "/todoList/": {
            "get": {
                "description": "Find all todo lists",
                "tags": [
                    "todoList"
                ],
                "summary": "FindTodoLists",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.TodoListDTO"
                            }
                        }
                    }
                }
            }
        },
        "/todoList/{name}": {
            "get": {
                "description": "Find a todo list by Name",
                "tags": [
                    "todoList"
                ],
                "summary": "FindTodoListByName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo List Name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TodoListDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a todo list by Name",
                "tags": [
                    "todoList"
                ],
                "summary": "DeleteTodoListByName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo List Name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TodoListDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "dtos.ItemDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duedate": {
                    "type": "string"
                }
            }
        },
        "dtos.TodoListDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ItemDTO"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}